requirements.txt
openai
streamlit
python-dotenv
fastapi
uvicorn
requests
components
streamlit-javascript
python-multipart
pydantic
streamlit
pydantic-settings